+++ model/client/account/account_transactions_response.go
@@ -2,7 +2,7 @@ package account
 
 import (
 	. "github.com/xyield/xrpl-go/model/client/common"
-	. "github.com/xyield/xrpl-go/model/transactions"
+	. "github.com/xyield/xrpl-go/model/transactions/types"
 )
 
 type AccountTransactionsResponse struct {
