+++ model/ledger/account_root.go
@@ -1,7 +1,7 @@
 package ledger
 
 import (
-	. "github.com/xyield/xrpl-go/model/transactions"
+	. "github.com/xyield/xrpl-go/model/transactions/types"
 )
 
 type AccountRoot struct {
