+++ model/ledger/signer_entry.go
@@ -1,7 +1,7 @@
 package ledger
 
 import (
-	. "github.com/xyield/xrpl-go/model/transactions"
+	. "github.com/xyield/xrpl-go/model/transactions/types"
 )
 
 type SignerEntry struct {
