+++ model/transactions/account_delete.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type AccountDelete struct {
 	BaseTx
 	Destination    Address
