+++ model/transactions/check_cancel.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type CheckCancel struct {
 	BaseTx
 	CheckID Hash256
