+++ model/transactions/check_cash.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type CheckCash struct {
 	BaseTx
 	CheckID    Hash256
