+++ model/transactions/check_create.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type CheckCreate struct {
 	BaseTx
 	Destination    Address
