+++ model/transactions/escrow_finish.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type EscrowFinish struct {
 	BaseTx
 	Owner         Address
