+++ model/transactions/nftoken_accept_offer.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type NFTokenAcceptOffer struct {
 	BaseTx
 	NFTokenSellOffer Hash256        `json:",omitempty"`
