+++ model/transactions/nftoken_burn.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type NFTokenBurn struct {
 	BaseTx
 	NFTokenID NFTokenID
