+++ model/transactions/nftoken_cancel_offer.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type NFTokenCancelOffer struct {
 	BaseTx
 	NFTokenOffer []Hash256
