+++ model/transactions/nftoken_create_offer.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type NFTokenCreateOffer struct {
 	BaseTx
 	Owner       Address `json:",omitempty"`
