+++ model/transactions/nftoken_mint.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type NFTokenMint struct {
 	BaseTx
 	NFTokenTaxon uint
