+++ model/transactions/path_step.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type PathStep struct {
 	Account  Address `json:"account,omitempty"`
 	Currency string  `json:"currency,omitempty"`
