+++ model/transactions/payment.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type Payment struct {
 	BaseTx
 	Amount         CurrencyAmount
