+++ model/transactions/payment_channel_claim.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type PaymentChannelClaim struct {
 	BaseTx
 	Channel   Hash256
