+++ model/transactions/payment_channel_fund.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type PaymentChannelFund struct {
 	BaseTx
 	Channel    Hash256
