+++ model/transactions/set_regular_key.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type SetRegularKey struct {
 	BaseTx
 	RegularKey Address `json:",omitempty"`
