+++ model/transactions/signer.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type Signer struct {
 	Account       Address
 	TxnSignature  []byte
