+++ model/transactions/signer_list_set.go
@@ -1,14 +1,13 @@
 package transactions
 
 import (
-	// TODO restructure to fix import cycle
-	"github.com/xyield/xrpl-go/model/ledger"
+	. "github.com/xyield/xrpl-go/model/ledger"
 )
 
 type SignerListSet struct {
 	BaseTx
 	SignerQuorum  uint
-	SignerEntries []ledger.SignerEntry
+	SignerEntries []SignerEntry
 }
 
 func (*SignerListSet) TxType() TxType {
