+++ model/transactions/trust_set.go
@@ -1,5 +1,9 @@
 package transactions
 
+import (
+	. "github.com/xyield/xrpl-go/model/transactions/types"
+)
+
 type TrustSet struct {
 	BaseTx
 	LimitAmount CurrencyAmount
