+++ model/transactions/types/address.go
@@ -1,0 +1,3 @@
+package types
+
+type Address string
