+++ model/transactions/types/currency_amount.go
@@ -1,4 +1,4 @@
-package transactions
+package types
 
 // TODO XRP vs issued currency marshal/unmarshal
 type CurrencyAmount struct {
