+++ model/transactions/types/hash128.go
@@ -1,0 +1,3 @@
+package types
+
+type Hash128 []byte
