+++ model/transactions/types/hash256.go
@@ -1,0 +1,3 @@
+package types
+
+type Hash256 []byte
