+++ model/transactions/types/nftoken_id.go
@@ -1,0 +1,3 @@
+package types
+
+type NFTokenID Hash256
