+++ model/transactions/types/nftoken_uri.go
@@ -1,0 +1,3 @@
+package types
+
+type NFTokenURI string
