+++ model/transactions/types/xaddress.go
@@ -1,0 +1,3 @@
+package types
+
+type XAddress string
