+++ model/transactions/types/xrp_currency_amount.go
@@ -1,0 +1,3 @@
+package types
+
+type XrpCurrencyAmount uint64
